# CLN/PERF: move RangeIndex._cached_data to RangeIndex._cache