The ``._cached_data`` attribute is not necessary. It was originally added to allow a check to see it the ``._data`` ndarray had been created, but that's also possible to do by a ``"_data" in _cache`` check in the new implemention, which IMO would be more idiomatic.

The new implementation has the benefit that the ``_data`` will be copied to new instances, saving the need to create a new ndarray for each new copy of the RangeIndex.

```python
>>> idx = pd.RangeIndex(1_000_000)
>>> idx[[1, 4]]  # this accesses ._data and saves it in cached_data (master) or _cache["_data"](this PR)
>> %timeit idx._shallow_copy()[[1, 4]]
2.55 ms ± 69.3 µs per loop  # master
17.7 µs ± 405 ns per loop  # this PR
```

xref #26565.